Backup 4
Modified Mob to spawn in a continous trail rather than a box.

Bug Fix: Flame thrower bullets were throwing uninstanated object errors
when colliding with cover_barriers. Caused by a deprecated script on
flame thrower bullet entity that was disabled but not removed, so it was
recieving the OnTriggerEnter calls.